budget_tracker_api/api/views/users.py
from django.contrib.auth import get_user_model
from rest_framework import viewsets
from rest_framework.permissions import IsAuthenticated

from budget_tracker_api.api.serializers.users import UserSerializer

UserModel = get_user_model()


class UserViewSet(viewsets.ModelViewSet):
    permission_classes = (IsAuthenticated,)
    serializer_class = UserSerializer
    queryset = UserModel.objects.all()
